move full height skeleton to components folder
