+++ src/views/components/index.tsx
@@ -1,3 +1,4 @@
 export { default as Loading } from './loading'
 export { default as Text } from './text'
 export { default as QR } from './qr'
+export { default as FullHeightSkeleton } from './full_height_skeleton'
